requirements.txt
Django>=5.2.7
djangorestframework>=3.16.1
django-cors-headers>=4.9.0
mysqlclient>=2.2.7
requests>=2.32.5
Pillow>=12.0.0
python-dotenv>=1.1.1
gunicorn>=23.0.0